+++ src/main/java/divinerpg/items/vanilla/ItemScythe.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class ItemScythe extends ItemModRanged {
     public ItemScythe() {super(Rarity.COMMON, "shooter_bullet", BulletType.SCYTHE_SHOT, SoundRegistry.DEEP_LAUGH, SoundSource.PLAYERS, -1, 0, null, 0);}
     @Override protected void spawnEntity(Level world, Player player, ItemStack stack, BulletType bulletType, String entityType) {super.spawnEntity(world, player, stack, isJackoman(player) ? BulletType.MEGA_SCYTHE_SHOT : BulletType.SCYTHE_SHOT, null);}
-    @Override public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {if(Minecraft.getInstance().player != null) tooltip.add(LocalizeUtils.rangedDam(isJackoman(Minecraft.getInstance().player) ? BulletType.MEGA_SCYTHE_SHOT.getDamage() : BulletType.SCYTHE_SHOT.getDamage()));}
+    @Override public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {if(Minecraft.getInstance().player != null) tooltip.add(LocalizeUtils.rangedDam(isJackoman(Minecraft.getInstance().player) ? (int)BulletType.MEGA_SCYTHE_SHOT.getDamage() : (int)BulletType.SCYTHE_SHOT.getDamage()));}
     private boolean isJackoman(Player player) {
         ItemStack stackBoots = player.inventory.armor.get(0);
         ItemStack stackLegs = player.inventory.armor.get(1);
